sync

+++ go/weed/weed_server/filer_server_handlers.go
@@ -130,7 +130,11 @@ func (fs *FilerServer) PostHandler(w http.ResponseWriter, r *http.Request) {
 	if replication == "" {
 		replication = fs.defaultReplication
 	}
-	assignResult, ae := operation.Assign(fs.master, 1, replication, fs.collection, query.Get("ttl"))
+	collection := query.Get("collection")
+	if collection == "" {
+		collection = fs.collection
+	}
+	assignResult, ae := operation.Assign(fs.master, 1, replication, collection, query.Get("ttl"))
 	if ae != nil {
 		glog.V(0).Infoln("failing to assign a file id", ae.Error())
 		writeJsonError(w, r, http.StatusInternalServerError, ae)
@@ -209,7 +213,7 @@ func (fs *FilerServer) DeleteHandler(w http.ResponseWriter, r *http.Request) {
 		err = fs.filer.DeleteDirectory(r.URL.Path, isRecursive)
 	} else {
 		fid, err = fs.filer.DeleteFile(r.URL.Path)
-		if err == nil {
+		if err == nil && fid != "" {
 			err = operation.DeleteFile(fs.master, fid, fs.jwt(fid))
 		}
 	}
